+++ crates/c-api/include/wasmtime/helpers.hh
@@ -91,4 +91,69 @@ public:                                                                        \
    */                                                                          \
   bool operator!=(const name &other) const { return !(*this == other); }
 
+#define WASMTIME_REF_WRAPPER(name, capi_type)                                  \
+public:                                                                        \
+  /**                                                                          \
+   * \brief Non-owning reference to an instance of this type.                  \
+   */                                                                          \
+  using Raw = capi_type##_t;                                                   \
+                                                                               \
+private:                                                                       \
+  Raw raw;                                                                     \
+                                                                               \
+public:                                                                        \
+  /**                                                                          \
+   * \brief Takes ownership of `val` and wraps it with this class.             \
+   */                                                                          \
+  explicit name(Raw raw) : raw(raw) {}                                         \
+  name(const name &other) { capi_type##_clone(&other.raw, &raw); }             \
+                                                                               \
+  name &operator=(const name &other) {                                         \
+    capi_type##_unroot(&raw);                                                  \
+    capi_type##_clone(&other.raw, &raw);                                       \
+    return *this;                                                              \
+  }                                                                            \
+  name(name &&other) {                                                         \
+    raw = other.raw;                                                           \
+    capi_type##_set_null(&other.raw);                                          \
+  }                                                                            \
+  name &operator=(name &&other) {                                              \
+    capi_type##_unroot(&raw);                                                  \
+    raw = other.raw;                                                           \
+    capi_type##_set_null(&other.raw);                                          \
+    return *this;                                                              \
+  }                                                                            \
+                                                                               \
+  ~name() { capi_type##_unroot(&raw); }                                        \
+                                                                               \
+  /**                                                                          \
+   * \brief Returns the underlying C API pointer.                              \
+   */                                                                          \
+  const Raw *capi() const { return &raw; }                                     \
+                                                                               \
+  /**                                                                          \
+   * \brief Returns the underlying C API pointer.                              \
+   */                                                                          \
+  Raw *capi() { return &raw; }
+
+#define WASMTIME_TOP_REF_WRAPPER(name, capi_type)                              \
+  WASMTIME_REF_WRAPPER(name, capi_type)                                        \
+                                                                               \
+  /**                                                                          \
+   * \brief Consumes ownership of the underlying `wasmtime_externref_t`        \
+   * and returns the result of `wasmtime_externref_to_raw`.                    \
+   */                                                                          \
+  uint32_t take_raw(Store::Context cx) {                                       \
+    uint32_t ret = capi_type##_to_raw(cx.capi(), &raw);                        \
+    capi_type##_set_null(&raw);                                                \
+    return ret;                                                                \
+  }                                                                            \
+                                                                               \
+  /**                                                                          \
+   * \brief Returns `wasmtime_externref_to_raw`.                               \
+   */                                                                          \
+  uint32_t borrow_raw(Store::Context cx) const {                               \
+    return capi_type##_to_raw(cx.capi(), &raw);                                \
+  }
+
 #endif // WASMTIME_HELPERS_HH
