+++ crates/c-api/include/wasmtime/types/tag.hh
@@ -19,7 +19,6 @@ namespace wasmtime {
  */
 class TagType {
   friend class Tag;
-  friend class Exn;
 
   struct deleter {
     void operator()(wasm_tagtype_t *p) const { wasm_tagtype_delete(p); }
