+++ crates/c-api/src/types.rs
@@ -33,3 +33,12 @@ pub use self::memory::*;
 pub use self::table::*;
 pub use self::tag::*;
 pub use self::val::*;
+
+#[cfg(feature = "gc")]
+mod arrayref;
+#[cfg(feature = "gc")]
+mod structref;
+#[cfg(feature = "gc")]
+pub use self::arrayref::*;
+#[cfg(feature = "gc")]
+pub use self::structref::*;
